# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartSpend API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions (filterable)
      parameters:
        - name: month
          in: query
          schema:
            type: string
          description: "YYYY-MM format"
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List budgets with spending progress
      parameters:
        - name: month
          in: query
          schema:
            type: string
          description: "YYYY-MM format, defaults to current month"
      responses:
        "200":
          description: Budgets with progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetWithProgress"

    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create or replace a monthly budget for a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories (predefined + custom)
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a custom category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a custom category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        note:
          type: ["string", "null"]
        date:
          type: string
          format: date
        isRecurring:
          type: boolean
        recurringPeriod:
          type: ["string", "null"]
          enum: [daily, weekly, monthly, "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amount
        - type
        - category
        - date
        - isRecurring
        - note
        - recurringPeriod
        - createdAt
    TransactionInput:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
          minLength: 1
        note:
          type: string
        date:
          type: string
          format: date
        isRecurring:
          type: boolean
        recurringPeriod:
          type: string
          enum: [daily, weekly, monthly]
      required:
        - amount
        - type
        - category
        - date
    TransactionUpdate:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        note:
          type: string
        date:
          type: string
          format: date
        isRecurring:
          type: boolean
        recurringPeriod:
          type: string
          enum: [daily, weekly, monthly]
    Budget:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        category:
          type: string
        amount:
          type: number
        month:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - category
        - amount
        - month
        - createdAt
    BudgetWithProgress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        category:
          type: string
        amount:
          type: number
        month:
          type: string
        spent:
          type: number
        percentage:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - category
        - amount
        - month
        - spent
        - percentage
        - createdAt
    BudgetInput:
      type: object
      properties:
        category:
          type: string
          minLength: 1
        amount:
          type: number
          minimum: 1
        month:
          type: string
          description: "YYYY-MM"
      required:
        - category
        - amount
        - month
    BudgetUpdate:
      type: object
      properties:
        amount:
          type: number
          minimum: 1
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        isDefault:
          type: boolean
        color:
          type: string
      required:
        - id
        - name
        - icon
        - isDefault
        - color
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        color:
          type: string
      required:
        - name
...